fix: remove duplicate compact functions from proxy-handler, use compact.ts module
The compact-related functions (isCompactRequest, estimateTokens, getCompactConfig,
getModelContextWindow, replaceCompactPrompt) were extracted to compact.ts but the
old local definitions in proxy-handler.ts were never removed, causing import conflicts
and ReferenceError at runtime. Also fix admin/providers.ts formatting.
